refactor(stella-context): split the three files crowding the 1500-line ceiling
`retrieval.rs` (1446), `writeback.rs` (1416) and `retrieval/tests.rs` (1416)
each sat within 90 lines of the gate's ratchet with no baseline entry, so the
next non-trivial addition to any of them would have failed `make gate`
outright — and the baseline accepts no new entries, so there was no
grandfathering to fall back on. Split along the seams each file already had,
following the `ranking`/`evidence`/`scope` fan-out the crate established:

- `retrieval/tuning.rs` — `RecallTuning` and every `DEFAULT_*` knob, with the
  reasoning for each value. What a recall is tuned by, separate from what it
  does.
- `retrieval/outcome.rs` — what a recall reports: `RecallResult`,
  `DroppedFrame`/`DropReason`, `SelectionReason`, `RecallTier`, and the
  crate-private `Ranked`.
- `retrieval/recall_tests.rs` — the end-to-end half of the retrieval tests, a
  sibling module like `evidence_tests`, since it shares nothing with the pure
  ranking units but `super::*`.
- `writeback/tests.rs` — the inline test module, moved out under the same
  pattern `store.rs` and `retrieval.rs` already follow.

Pure moves: every public path is preserved by re-export, and the crate's test
count is unchanged at 167 before and after. `retrieval.rs` is now 958 lines
and `writeback.rs` 997.

The crate README's "God files" section carried a hard-coded "1498 lines"
claim about `src/store/tests.rs`, which is 1117 — the exact drift AGENTS.md
warns a number-in-prose causes. It is replaced with a pointer to the seams
and no line count at all; the baseline and the guard stay the only copies.

Closes #3705
